\name{plotting}
\alias{splot}
\alias{splot2}

\title{ Plotting model results }
\description{
  Function to make a plot of sites, ....
}
\usage{
splot(x, ...)
splot2(x, ...)
}
\arguments{
    \item{x}{ A simulation result object }  
    \item{...}{ Additional arguments for plot() }  	
}
\details{
  
}
\value{
  None. Only a side effect (the plot)
}
\author{ Robert J. Hijmans }

\examples{
}
\keyword{ spatial }
