\name{plotting}
\alias{splot}

\title{ Plotting model results }
\description{
  Function to make a plot of sites, ....
}
\usage{
splot(x)
}
\arguments{
    \item{x}{ A simulation result object }  
}
\details{
  
}
\value{
  None. Only a side effect (the plot)
}
\author{ Robert J. Hijmans }

\examples{
}
\keyword{ spatial }
